Multilevel selection working correctly now with the component abstraction
TODO:
Need to figure out how to actively update the bounds and position of
the component object and also also figure out how to update the rest of
the parameters as the features are being moved around. Perhaps it has
to be done during the movement phase or the first rendering phase.

This yet to be fixed !!! Will create an issue on this later when the
move tool and everything is completed. For now its just multilevel
selection.

Also reformatted, cleaned and added documentation to modified methods