All the layers seem to be changing correctly
TODO: Need to fix the component placement business

Copies the edge features onto every new layer that is being created

Colors are working correctly and I took out all the extra console logs

Updated the multilayer position tools to properly index and place the components in the correct layers

Delete is working correctly

TODO:  I need to change the ui generation to be a method that does all
the ui generation every time, the action happens rather than modifying
the deltas

Updated what the variable means

TODO: Fix weird UI issues in the labels and selection colors. Do this
with knockoutjs and simplify this maddness.