Updated connection termination conditions
Updated the data model of connections and also implemented the
termination condition when connecting to other connections

Fixed issue when connection object was not being deleted

Segment generation working correctly (for now)

TODO: Need to rework the rotation algorithm to find the nearest segment. Currently the whole system still relies on 'start' and 'end' waypoints that need to be removed.