Improved stability of valve insertion
Basically there was an error in how we were splitting up the segment into 2 pieces. The logic was incorrect in how we decided which break point was closer to the segment ends. New method doesn't just compute distance to one point but instead computes total distance of the configs and sees which segment configs have the least length.