Adding Anode and Cathode
Anode and Cathode positions for electrophoresis on  a chip.
Both have 'pegs' which I intend to be placeholders for a wire, or perhaps a scaffold over which a wire could be looped. They are based off of ports since the wire will likely be coming in from the control layer.