Create fabricate_chip1.js
This is our attempt at the extra credit. Since we have three microfluidic chips, and the assignment asks for the fabrication protocol of the chip design, we figured it would be sufficient to only have one microfluidic chip (out of three of ours) to show our attempt at the extra credit. We hope that this is enough (and isn't us skimping out on the effort!)