Add gel channel
This gel channe for separation by gel electrophoresis l is based on CellTrapL. Its my attempt to make a channel that is more adjustable with two height parameters, one for the main channel and one for some "reservoir" channels on its sides. I'm not sure how functional it will be for actual use, but I hoped it would make it easier to cast gels. You can make elevated regions where air bubbles might be able to escape using this component.  It may be helpful when you are pumping in a viscous fluid. By having multiple heights there could be more room for the gel to expand without changing the dimensions of the inlet. Might end up doing nothing but could be fun to play with.