Gelchannel (#178)
* Add electrophoresis to view manager

link event handler to ux

* Update viewManager.js

* Add gel channel

This gel channe for separation by gel electrophoresis l is based on CellTrapL. Its my attempt to make a channel that is more adjustable with two height parameters, one for the main channel and one for some "reservoir" channels on its sides. I'm not sure how functional it will be for actual use, but I hoped it would make it easier to cast gels. You can make elevated regions where air bubbles might be able to escape using this component.  It may be helpful when you are pumping in a viscous fluid. By having multiple heights there could be more room for the gel to expand without changing the dimensions of the inlet. Might end up doing nothing but could be fun to play with.

* add gelchannel to featureset

copied the code for celltrapL and modify for gel channel

* add gelchannel ui element

copy celltrapL ui and insert gelchannel

* add gelchannel to toolbar

register the event handlers

* gelchannel tool association to viewmanager

instantiate placement tool

* Correction of an accidental copy paste

There was some duplicate code I accidentally copied in a previous commit (it still worked). Fixed it in this one